refactor(speech): extract macos26 analyzer tests to sibling file
Move the single inline `#[cfg(test)] mod tests` out of
`primer-speech/src/macos26/analyzer.rs` into a sibling
`macos26/analyzer/tests.rs` (non-root module → one dir deeper, mkdir).

Pure verbatim test-module relocation: 721 → 308 lines (−413). The test
module is a descendant, so `use super::*` reaches all parent privates
with no visibility edits; the body carries no depth-sensitive paths
(no include_str!/"../"/env!), so the deeper dir is irrelevant.

Guarded under `--features macos-native-26` (swiftc sidecar); 11 analyzer
tests preserved (75 lib tests total, identical pre/post).
